Rework rendering to fix paging
(Hopefully for the last time.)

Change summary:
- fix paging bug where extra lines were being printed
- fmt() now returns a list of lines instead of printing them out
- rendering step split into two - calculating the page to be printed and then actually printing it
- no more rendering work during the paging step \o/
- added basic backward navigation
- reordered the input handling stuff

Explanation:

This was an interesting bug to fix. Basically, the paging logic expected that each screenful be a 1:1 mapping of screen rows to lines, i.e. n(rows) = n(lines). However some lines are wider than the screen permits and this was handled by, but this made text wrap, and so more lines than the screen height got printed, ie n(rows) < n(lines). This made the pager print more than a screenful at a time.

Fix was to split the rendering into two stages. The first stage, get_render_body(), splits the raw text into lines and adds formatting. (This also has the side benefit of eliminating the janky render stuff that previously happened in the paging phase and putting it in the _actual_ render phase). The second stage, page(), now does the paging work. This way, no unexpected extra lines get created _during_ the paging step, so the screenful logic works correctly all of the time.

Also added a bit of basic navigation logic, but it's untested for now.

Reordered the input handling as well, since I'd accidentally broken it and not noticed.