Major refactor Change summary: - refactored state variable into a Browser class - (the Browser class now gets and renders documents instead of `get_document_ez` and `render` doing those things) - added a basic config file loader, just reads a JSON and sets some options - added instructions to the readme - started to add type annotations to stuff - remove redundant getch()
Got a code review from the folks in #python on libera.chat. Changes they suggested:
1) remove the redundant `global` statements (I'm not sure why they were there in the first place.)
2) refactor state into a class
3) remove parentheses around if conditions
4) use the with...as syntax to ensure destruction of the socket fp
5) add type annotations

Number 2 is the focus of this commit. Using a ball of global state is an awful idea. So I turned it into an object and while there are a bunch of globals still it's quite a bit better. Type annotations are also cool, they make it so VSCode no longer screams at me for setting debug_logger to None even though I specifically check for None before calling it.

Also removed a getch() method left over from the early days of the pager.