MORE renderer fixes
Lines that didn't start with one of the gemtext line formats were getting skipped if they weren't links / in a preformatted block. I realised I couldn't use if...elif...else because some lines might meet multiple conditions.

Switched back to if...if...if. However this uses that ugly appended variable, more thinking is required to refactor this a bit more elegantly.